* src/db-artwork-writer.c (ipod_buffer_grow_file): printf(3) type 	  warnings caused compile error with FreeBSD 6.0-stable/GCC 	  3.4.4. Thanks to Mike Heffner for the patch.
git-svn-id: https://gtkpod.svn.sourceforge.net/svnroot/gtkpod/libgpod/trunk@1309 f01d2545-417e-4e96-918e-98f8d0dbbcb6

+++ src/db-artwork-writer.c
@@ -98,14 +98,14 @@ ipod_buffer_grow_file (struct iPodMmapBuffer *mmap_buf, off_t new_size)
 	result = lseek (mmap_buf->fd, new_size, SEEK_SET);
 	if (result == (off_t)-1) {
 		g_print ("Failed to grow file to %lu: %s\n", 
-			 new_size, strerror (errno));
+                        (unsigned long)new_size, strerror (errno));
 		return -1;
 	}
 	result = 0;
 	result = write (mmap_buf->fd, &result, 1);
 	if (result != 1) {
 		g_print ("Failed to write a byte at %lu: %s\n", 
-			 new_size, strerror (errno));
+                        (unsigned long)new_size, strerror (errno));
 		return -1;
 	}
